fix some bugs

+++ admin/views/tsjs/view.html.php
@@ -13,16 +13,10 @@ JLoader::register('TSJsHelper', JPATH_COMPONENT.'/helpers/tsjs.php');
  */
 class TSJViewTSJs extends JViewAbstract
 {
-	//public $layout='';
-
-	//список записей
 	protected $items;
-	//объект постраничной навигации
 	protected $pagination;
-
-	protected $dbuser;
-	
-	public $param;
+	protected $state;
+    public $dbuser;
 
 	/**
 	 * TSJs view display method
@@ -31,66 +25,53 @@ class TSJViewTSJs extends JViewAbstract
 	function display($tpl = null)
 	{
 		// Get data from the model
-		$form = $this->get('Form');
-		$item = $this->get('Item');
-		$state = $this->get('State');
-		$pagination = $this->get('Pagination');
-		$this->param = $this->get('Params');
-		
+		$this->form = $this->get('Form');
+		$this->state = $this->get('State');
+		$this->pagination = $this->get('Pagination');
+		$this->param = $this->get('Params');      
+        
 		// Check for errors.
 		if (count($errors = $this->get('Errors')))
 		{
 			JError::raiseError(500, implode('<br />', $errors));
 			return false;
 		}
-
-		// Assign the Data
-		$this->form = $form;
-		$this->item = $item;
-		$this->state = $state;
-
-		TSJsHelper::addSubmenu('tsjs');
 				
 		// Get data from the model
 		if ($this->getLayout() == 'city')
 		{
-			$this->cityitems = $this->get('CityItems');
+			$this->cityitems = $this->get('Items');
 			$this->setLayout('city');
 			//$layout = 'city';
 		}
 		else if ($this->getLayout() == 'street')
 		{
-			$this->streetitems = $this->get('StreetItems');
+			$this->streetitems = $this->get('Items');
 			$this->setLayout('street');
 			//$layout = 'street';
 		}
 		else if ($this->getLayout() == 'address')
 		{
-			$this->addressitems = $this->get('AddressItems');
+			$this->addressitems = $this->get('Items');
 			$this->setLayout('address');
 			//$layout = 'address';
 		}
 		else if ($this->getLayout() == 'account')
 		{
-			$this->accountitems = $this->get('AccountItems');
+			$this->accountitems = $this->get('Items');
 			$this->setLayout('account');
 			//$layout = 'account';
 		}
 		else
 		{
-			$this->items = $this->get('Items');
-			$this->setLayout('tsjs');
-			//$layout = 'tsjs';
-			//$this->items = $this->get('Item');
-			//$this->setLayout('tsjs');
+            //JFactory::getApplication()->enqueueMessage('Debug: TSJViewTSJs::display layout=default');
+			$this->setLayout('default');
+            $this->defaultitems = $this->get('Items');
+            //JFactory::getApplication()->enqueueMessage('Debug: TSJViewTSJs::display $this->defaultitems='.$this->defaultitems);
 				
 		}
-
-		//$this->sortDirection = $state->get('filter_order_Dir');
-		//$this->sortColumn = $state->get('filter_order');
-
-		$this->pagination = $pagination;
-
+        
+		TSJsHelper::addSubmenu('tsjs');
 		// Set the toolbar
 		$this->addToolBar();
 
@@ -103,7 +84,7 @@ class TSJViewTSJs extends JViewAbstract
 		parent::display($tpl);
 
 		// Set the document
-		$this->setDocument();
+		//$this->setDocument();
 	}
 
 	/**
@@ -112,16 +93,18 @@ class TSJViewTSJs extends JViewAbstract
 	protected function addToolBar()
 	{
 		$state	= $this->get('State');
-		//$canDo	= TSJsHelper::getActions('com_tsjs', 'category', $state->get('filter.category_id'));
+		$user	= JFactory::getUser();
+        
+		//$canDo	= TSJsHelper::getActions('com_tsj', 'category', $state->get('filter.category_id'));
         if (version_compare(JPlatform::RELEASE, '12', '<')){
-            //$canDo	= TSJsHelper::getActions('com_tsjs');
+            $canDo	= TSJsHelper::getActions('com_tsj');
         }
         else{
             $categoryId	= $this->state->get('filter.category_id');
-            //$canDo = JHelperContent::getActions('com_tsjs');
+            $canDo	= JHelperContent::getActions('com_tsj');
         }
         
-		$user	= JFactory::getUser();
+        //JFactory::getApplication()->enqueueMessage('Debug: TSJViewTSJs::addToolBar state='. $state. ' user='.$user);
 		
 		// Заголовок
 		JToolBarHelper::title(JText::_('COM_TSJ_MANAGER_TSJS'));
@@ -133,46 +116,63 @@ class TSJViewTSJs extends JViewAbstract
 		
 		if ($this->getLayout() == 'street')
 		{
-			JToolBarHelper::addNew('street.add');
-			//if ($canDo->get('core.edit'))			
+        	if ($canDo->get('core.create')) {
+                JToolBarHelper::addNew('street.add');
+            }
+			if ($canDo->get('core.edit'))			
 			{
 				JToolBarHelper::editList('street.edit');
 			}
-			//if ($canDo->get('core.remote'))			
+			if ($canDo->get('core.delete'))			
 			{
 				JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'street.remove');
 			}
 		}
 		else if ($this->getLayout() == 'city')
 		{
-			//if ($canDo->get('core.create')) {
+			if ($canDo->get('core.create')) {
 				JToolBarHelper::addNew('city.add');
-			//}
-			JToolBarHelper::editList('city.edit');
-			JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'city.remove');
+			}
+            if ($canDo->get('core.edit')){
+                JToolBarHelper::editList('city.edit');
+            }
+            if ($canDo->get('core.delete')){
+                JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'city.remove');
+            }
 		}
 		else if ($this->getLayout() == 'address')
 		{
-			JToolBarHelper::addNew('address.add');
-			JToolBarHelper::editList('address.edit');
-			JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'address.remove');
+			if ($canDo->get('core.create')) {        
+                JToolBarHelper::addNew('address.add');
+            }
+            if ($canDo->get('core.edit')){
+                JToolBarHelper::editList('address.edit');
+            }
+            if ($canDo->get('core.delete')){
+                JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'address.remove');
+            }
 		}
 		else if ($this->getLayout() == 'account')
 		{
-			JToolBarHelper::addNew('account.add');
-			JToolBarHelper::editList('account.edit');
-			JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'account.remove');
+        	if ($canDo->get('core.create')) {
+                JToolBarHelper::addNew('account.add');
+            }
+            if ($canDo->get('core.edit')){
+                JToolBarHelper::editList('account.edit');
+            }
+            if ($canDo->get('core.delete')){
+                JToolBarHelper::deleteList('Вы действительно хотите удалить выбранные записи ?', 'account.remove');
+            }
 		}
 		else{
-		//JToolBarHelper::divider();
-		$toolbar->loadButtonType('Import', true);		
-		
-		if (version_compare(JPlatform::RELEASE, '12', '<')){
-			$toolbar->appendButton('Import', 'tsjs-import', 'COM_TSJ_CONFIG_IMPORT', 'tsjs.import',false);
-		}
-		else{
-			$toolbar->appendButton('Import', 'import', 'tsjs-import', 'COM_TSJ_CONFIG_IMPORT', 'tsjs.import', false);		
-		}
+            $toolbar->loadButtonType('Import', true);		
+            
+            if (version_compare(JPlatform::RELEASE, '12', '<')){
+                $toolbar->appendButton('Import', 'tsjs-import', 'COM_TSJ_CONFIG_IMPORT', 'tsjs.import',false);
+            }
+            else{
+                $toolbar->appendButton('Import', 'import', 'tsjs-import', 'COM_TSJ_CONFIG_IMPORT', 'tsjs.import', false);		
+            }
 
 		$doc = JFactory::getDocument();
 		//$icon_48_import = " .icon-48-tsjs {background:url(components/com_tsj/images/header/icon-48-importer.png) no-repeat; }";
@@ -181,32 +181,18 @@ class TSJViewTSJs extends JViewAbstract
 		$doc->addStyleDeclaration($icon_32_import);
 		
 		}
-
-		// Add a batch button
-		/*if ($user->authorise('core.create', 'com_tsjs') && $user->authorise('core.edit', 'com_tsjs') && $user->authorise('core.edit.state', 'com_tsjs'))
-		{
-			JHtml::_('bootstrap.modal', 'collapseModal');
-			$title = JText::_('JTOOLBAR_BATCH');
-
-			// Instantiate a new JLayoutFile instance and render the batch button
-			$layout = new JLayoutFile('joomla.toolbar.batch');
-
-			$dhtml = $layout->render(array('title' => $title));
-			$toolbar->appendButton('Custom', $dhtml, 'batch');
-		}*/	
 		
 		// Options button.
 		if (JFactory::getUser()->authorise('core.admin', 'com_tsj'))
 		{
 			JToolBarHelper::preferences('com_tsj');
-			//JToolBarHelper::divider();
 		}
 
 		JToolBarHelper::help( 'Components_TSJ', true );		
 		
 		if (version_compare(JPlatform::RELEASE, '12', '>='))
 		{		
-			JHtmlSidebar::setAction('index.php?option=com_tsjs&view=tsjs');
+			JHtmlSidebar::setAction('index.php?option=com_tsj&view=default');
 		}
 	}
